pyproject.toml
[project]
name = "events-face"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt[crypto]>=5.5.0",
    "flake8>=7.2.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "ruff>=0.12.0",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
